src/main/java/com/acciente/securetodo/resources/TodoUserResource.java
/*
 * Copyright 2016 - 2017, Acciente LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.acciente.securetodo.resources;

import com.acciente.securetodo.api.TodoUser;
import com.acciente.securetodo.core.TodoUserService;

import javax.ws.rs.Consumes;
import javax.ws.rs.POST;
import javax.ws.rs.Path;
import javax.ws.rs.Produces;
import javax.ws.rs.core.MediaType;

@Path("/users")
@Consumes(MediaType.APPLICATION_JSON)
@Produces(MediaType.APPLICATION_JSON)
public class TodoUserResource {
   private final TodoUserService todoUserService;

   public TodoUserResource(TodoUserService todoUserService) {
      this.todoUserService = todoUserService;
   }

   @POST
   public TodoUser createUser(TodoUser todoUser) {
      return todoUserService.createUser(todoUser);
   }
}
